requirements.txt
autopep8==1.5
azure-common==1.1.24
azure-core==1.2.1
azure-nspkg==3.0.2
azure-storage==0.36.0
azure-storage-blob==12.1.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.8
idna==2.8
isodate==0.6.0
msrest==0.6.11
numpy==1.18.1
oauthlib==3.1.0
pandas==1.0.1
pycodestyle==2.5.0
pycparser==2.19
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
six==1.14.0
urllib3==1.25.8
